Add files via upload
this was a test on vs code using the AI agent, noticed that both cursor and vs code are pretty good at editing the files, but cursor gets a faster answer to the prompt. And Clearly using the same prompt on both vs and cursor , different stiles of animation come up , didi go with more of a pastel color game on the vs project instead of the cyberpunk feel on cursor, but both code are 80% the same, using same functions on javascript to create the game.